pyproject.toml
[project]
name = "department-portal"
version = "0.1.0"
description = ""
authors = [
    {name = "molumolu",email = "kan.pavel@ilcoud.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "boto3 (>=1.40.60,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false